requirements.txt
PyYAML==5.1.2
requests==2.22.0
confuse==1.0.0